+++ src/client/cypress/e2e/mainPage/import.cy.js
@@ -133,9 +133,9 @@ describe("Test for importing boreholes.", () => {
     cy.dataCy("import-borehole-button").click();
 
     // Select borehole csv file
-    getImportFileFromFixtures("boreholes-missing-fields-and-duplicates.csv", null)
+    getImportFileFromFixtures("boreholes-missing-required-fields.csv", null)
       .then(fileContent => {
-        const file = new File([fileContent], "boreholes-missing-fields-and-duplicates.csv", {
+        const file = new File([fileContent], "boreholes-missing-required-fields.csv", {
           type: "text/csv",
         });
         let boreholeFile = new DataTransfer();
@@ -153,9 +153,7 @@ describe("Test for importing boreholes.", () => {
       .should("contain", "Row1")
       .should("contain", "Field 'location_x' is required.")
       .should("contain", "Row2")
-      .should("contain", "Row4")
-      .should("contain", "Row5")
-      .should("contain", "Borehole with same Coordinates (+/- 2m) and same TotalDepth is provided multiple times.");
+      .should("not.contain", "Row3");
   });
 
   it("can select workgroup when importing boreholes", () => {